# docs/openapi.yaml
openapi: 3.0.3
info:
  title: ECE651 Travel Planner API
  version: 1.0.0
servers:
  - url: http://localhost:8080
paths:
  /api/trips:
    get:
      summary: List trips
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TripDto'
    post:
      summary: Create trip
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TripRequest'
      responses:
        '200':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TripDto'
        '400':
          description: Bad Request
  /api/trips/generate:
    post:
      summary: Generate trip draft with AI notes
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TripGenerateRequest'
      responses:
        '200':
          description: Generated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TripDto'
        '400':
          description: Bad Request
  /api/trips/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      summary: Get trip
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TripDto'
        '404':
          description: Not Found
    put:
      summary: Update trip
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TripRequest'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TripDto'
        '404':
          description: Not Found
    delete:
      summary: Delete trip
      responses:
        '204':
          description: No Content
        '404':
          description: Not Found
...
components:
  schemas:
    TripDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        titleOrDestination:
          type: string
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        travelers:
          type: string
        budget:
          type: string
        notes:
          type: string
        status:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    TripRequest:
      type: object
      required: [titleOrDestination, travelers, budget, status]
      properties:
        titleOrDestination:
          type: string
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        travelers:
          type: string
        budget:
          type: string
        notes:
          type: string
        status:
          type: string
    TripGenerateRequest:
      type: object
      required: [titleOrDestination, startDate, endDate, travelers, budget]
      properties:
        titleOrDestination:
          type: string
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        travelers:
          type: string
          enum: [SOLO, COUPLE, FAMILY, GROUP]
        budget:
          type: string
          enum: [BUDGET, MEDIUM, LUXURY]
        interests:
          type: array
          items:
            type: string
...